In the last article, add notes about stack alignment
I figured this out shortly after publishing the article. I've since
gotten a good handle it the concepts, so it's time to capture it in
writing.

Also shorten all __attribute__ to __attribute since that's how I've
decided to write it from now on. Why use a longer name when the shorter
is equally good?