Update libc review about the _Atomic qualifier
NRK points out that Clang's C11 atomic functions require the _Atomic
qulaifier, as specified in the standard. In other words, C11 atomics are
not powerful enough to precisely implement some atomic algorithms, and
this is a practical issue on a major implementation. This is the last
straw, and from now on I'm sticking to compiler intrisic extensions.